# [Schema Registry Avro] Add a serialize performance test