--- a/sdk/schemaregistry/perf-tests/schema-registry-avro/package.json
+++ b/sdk/schemaregistry/perf-tests/schema-registry-avro/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@azure-tests/perf-schema-registry-avro",
+  "sdk-type": "perf-test",
+  "version": "1.0.0",
+  "description": "",
+  "main": "",
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@azure/schema-registry": "^1.0.0",
+    "@azure/schema-registry-avro": "^1.0.0-beta.7",
+    "@azure/identity": "^2.0.1",
+    "@azure/test-utils-perf": "^1.0.0",
+    "dotenv": "^8.2.0"
+  },
+  "devDependencies": {
+    "@types/node": "^12.0.0",
+    "eslint": "^7.15.0",
+    "prettier": "^2.5.1",
+    "rimraf": "^3.0.0",
+    "tslib": "^2.2.0",
+    "ts-node": "^10.0.0",
+    "typescript": "~4.2.0"
+  },
+  "private": true,
+  "scripts": {
+    "perf-test:node": "ts-node test/index.spec.ts",
+    "audit": "node ../../../common/scripts/rush-audit.js && rimraf node_modules package-lock.json && npm i --package-lock-only 2>&1 && npm audit",
+    "build": "npm run clean && tsc -p .",
+    "build:samples": "echo skipped",
+    "build:test": "echo skipped",
+    "check-format": "prettier --list-different --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"test/**/*.ts\" \"*.{js,json}\"",
+    "clean": "rimraf dist dist-* types *.tgz *.log",
+    "format": "prettier --write --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"test/**/*.ts\" \"*.{js,json}\"",
+    "integration-test:browser": "echo skipped",
+    "integration-test:node": "echo skipped",
+    "integration-test": "echo skipped",
+    "lint:fix": "eslint --no-eslintrc -c ../../../.eslintrc.internal.json package.json test --ext .ts --fix --fix-type [problem,suggestion]",
+    "lint": "eslint --no-eslintrc -c ../../../.eslintrc.internal.json package.json test --ext .ts",
+    "pack": "npm pack 2>&1",
+    "unit-test:browser": "echo skipped",
+    "unit-test:node": "echo skipped",
+    "unit-test": "echo skipped",
+    "test:browser": "echo skipped",
+    "test:node": "echo skipped",
+    "test": "echo skipped"
+  }
+}