--- a/sdk/schemaregistry/perf-tests/schema-registry-avro/sample.env
+++ b/sdk/schemaregistry/perf-tests/schema-registry-avro/sample.env
@@ -1,0 +1,20 @@
+SCHEMA_REGISTRY_ENDPOINT=<Endpoint URL>
+SCHEMA_REGISTRY_GROUP=<Group name for schemas in registry>
+
+# Used to authenticate using Azure AD as a service principal for role-based authentication
+# in the tokenAuth sample.
+#
+# See the documentation for `EnvironmentCredential` at the following link:
+# https://docs.microsoft.com/javascript/api/@azure/identity/environmentcredential
+AZURE_TENANT_ID=<AD tenant id or name>
+AZURE_CLIENT_ID=<ID of the user/service principal to authenticate as>
+AZURE_CLIENT_SECRET=<client secret used to authenticate to Azure AD>
+
+# Used in samples that use Event Hubs. Retrieve these values from an Event Hub in the Azure portal.
+EVENTHUB_CONNECTION_STRING=<Event Hub connection string>
+EVENTHUB_NAME=<Event Hub name>
+CONSUMER_GROUP_NAME=<Event Hub Group name>
+
+# Used in samples that use Event Grid. Retrieve these values from an Event Grid topic in the Azure portal
+EVENT_GRID_TOPIC_ENDPOINT=<Event Grid topic endpoint>
+EVENT_GRID_TOPIC_API_KEY=<Event Grid topic API key>