--- a/sdk/schemaregistry/perf-tests/schema-registry-avro/test/index.spec.ts
+++ b/sdk/schemaregistry/perf-tests/schema-registry-avro/test/index.spec.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { PerfProgram, selectPerfTest } from "@azure/test-utils-perf";
+import { SerializeTest } from "./serialize.spec";
+
+import dotenv from "dotenv";
+dotenv.config();
+
+console.log("=== Starting the perf test ===");
+
+const perfProgram = new PerfProgram(selectPerfTest([SerializeTest]));
+
+perfProgram.run();